stilya_project/requirements.txt
# Azure SDK and Cloud Services
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
azure-servicebus>=7.11.0

# FastAPI and Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML Libraries
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
openai>=1.6.0
langchain>=0.1.0
langchain-community>=0.0.12

# Vector Search and Databases
faiss-cpu>=1.7.4
chromadb>=0.4.20
psycopg2-binary>=2.9.9
redis>=5.0.1

# Computer Vision
opencv-python>=4.8.1.78
Pillow>=10.1.0

# Data Processing
numpy>=1.24.3
pandas>=2.1.4
scikit-learn>=1.3.2

# Monitoring and Observability
structlog>=23.2.0
azure-monitor-opentelemetry>=1.2.0
prometheus-client>=0.19.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Message Queue (if not using Azure Service Bus)
kafka-python>=2.0.2

# MLOps (Optional)
mlflow>=2.8.1
prefect>=2.14.10

# Async support
aiohttp>=3.9.0
asyncpg>=0.29.0